memory/mem_encrypt: report failure and quit test when module failed to compile
it would silently pass for the current code when module failed to build,
and men_encryt=on is in kernel cmdline before the test starts.

https://beaker.engineering.redhat.com/recipes/18574663#task194109866

Fix it by adding the rlPhaseEnd which is a point for reporting. And make
the test quit early when module build failed.

+++ general/memory/function/mem_encrypt/runtest.sh
@@ -50,6 +50,14 @@ rlJournalStart
 		rlJournalEnd
 		exit 0
 	}
+	rpm -q "${kname}-devel-${kversion}-${krelease}" || rlRpmInstall "${kname}-devel" "$kversion" "$krelease" "$(uname -m)"
+	rpm -q "${kname}-devel-${kversion}-${krelease}" || rlDie "no ${kname}-devel package available"
+	if ! rlRun "make -C sme_module" 0; then
+		echo "Quit test as fail to build test module"
+		rlPhaseEnd
+		rlJournalEnd
+		exit 0
+	fi
 	if test -f $firstboot && grep "done" $firstboot; then
 		grep "mem_encrypt=on" /proc/cmdline
 		rlRun "rm -f $firstboot"
@@ -86,15 +94,18 @@ rlJournalStart
 		rlJournalEnd
 		rstrnt-reboot
 	fi
-	rpm -q "${kname}-devel-${kversion}-${krelease}" || rlRpmInstall "${kname}-devel" "$kversion" "$krelease" "$(uname -m)"
-	rpm -q "${kname}-devel-${kversion}-${krelease}" || rlDie "no ${kname}-devel package available"
-	rlRun "make -C sme_module" 0
 	rlRun "dmesg -C"
 	rlRun "insmod sme_module/sme_test.ko" 1
 	rlRun "journalctl -k --no-hostname | grep SMEtest | awk '{print \$5,\$6,\$7,\$8,\$9,\$10,\$11,\$12}' | grep -v deadbeef" 0
 	rlRun "dmesg -C"
 	if [ -e $firstboot ]; then
-		grep skip_cleanup_cmdline $firstboot && echo "reserve mem_encrypt=on in cmdline" && rm -f $firstboot && exit 0
+		if grep skip_cleanup_cmdline $firstboot; then
+			echo "reserve mem_encrypt=on in cmdline"
+			rm -f $firstboot
+			rlPhaseEnd
+			rlJournalEnd
+			exit 0
+		fi
 		rlRun "grubby --remove-args mem_encrypt=on --update-kernel DEFAULT"
 		rlRun "echo done > $firstboot"
 		rlPhaseEnd
